Organize classes into multiple packages
