+++ src/main/java/boba/task/Deadline.java
@@ -1,4 +1,4 @@
-package boba;
+package boba.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
