+++ src/main/java/boba/task/Todo.java
@@ -1,4 +1,4 @@
-package boba;
+package boba.task;
 
 /**
  * Represents a todo task without any date/time attached.
