+++ src/main/java/boba/util/CheerLoader.java
@@ -1,4 +1,4 @@
-package boba;
+package boba.util;
 
 import java.io.BufferedReader;
 import java.io.IOException;
